Fix errors in init, generate and countgen
-clear the vector when running init again.  I think new generators were
getting tacked on to the end of the vector since the vector is a member
of the class and not getting completely renewed on every call to init
- max_val for count gen does not need a -1, its a modulo operation so
a 3 data bit slot has a max_val of 7, which is ok when 7 % 8, but not ok
when 7 % 7 (=0).  I might change the meaning of max val to be less
confusing
- I wasn't storing the modulo result of the mVal in the count gen, its
just a silly mistake.