Add runt pulse detection on frames and data
When advanced analysis is enabled, the analyzer will mark frame sync
and data signals with transitions that are not caught be the clock.
These would indicate an incorrect setting in the clock edge of analyzer
or a problem with the data source.

notes:
I had an approach which was intended to look ahead in the stream before
moving the frame and data channel locations to determine if a transition
would occur, and if two transitions would, then mark the frame.
However, this code did not work.  I have left it in in case someone
at some point in the future wants to debug it.  I will be removing all
traces of the failed code in the next commit.

I also noticed that the placement of markers would cause some visual
artifacts (if markers were placed on top of one another, when the markers
were scrolled into view, all markers on that channel would disappear.
There is some other commented out code which were parts of other failed
attempts at getting this feature working and marking the edge of the
transition that would get missed, instead of the middle of the clock
like they are currently marked.  I also opted to show data values
(0 or 1) only when the advanced analysis is off to make it easier to
spot problems in the data stream.  There is a limit to the number of
markers that can be shown on the screen, and with the numbers enabled,
the error markers would not show unless you were zoomed way in.

I think I want to get these marks into the frame bubbles as warnings
or errors instead of marking the channel.