Display short frames
- Only try to decode data that is in a regular size slot.
- Display error frame for short slot but do not attempt to decode the
value (how would you even?)
- Move result commit and update progress into the slot analysis since
thats where we are creating frames.

found my bug.  If the frame is short, then our starting and ending index
needs to account for the fact that the slot is short.  Also need to keep
the "zero" length check in place (for sample a 1 bit wide slot would try
to start and end on the same sample edge, which crashes the renderer.