Detect clock rate errors, add Frame Rate setting
- Add sample rate setting (will be used to calculate bitclock errors)
- Get sample rate and calculate the desired number of samples between
sample clocks, report an error if the difference is greater than 2
samples in either direction (e.g. if desired is 20, report an error if
the number of samples encountered is 18 or 22)
- Report the desired number of samples and the measured samples (if in
error) in the frame.mData2 slot (likely going to be removed in the
future).
- Add "E: " to all errors amd "W: " reported in frame_v2 for easier
searching
- Update the GetMinimumSampleRateHz function to return a correct
value depending on parameters entered into the analyzer
- Add the final BITCLOCK_ERROR flag (its the last flag we can add
to the mFlags value

Notes, I could not get proper sample rate values when mSampleRate was
a U32, no idea why (I reported the value returned in the mData2 as a
debug print option)