Fix off by one error in right justified data
I was off by one in my starting index calculation, which
caused the index to be very huge when the slot size and
the data size were the same (unsigned int that was set to
a negative number when slot size == data size).
This also caused the parsed data to be off by one as well, so
this fixed an error in interpreting the right justified data
too.  I think I ran into this bug during development, but I
couldn't get the debugger to work (They have since improved the
documentation) and just decided to move on at the time since I couldn't
easily track down the root cause of the error.  Now that I was
able to get the debugger to work, it was pretty easy to see
what was going wrong and fix it.